Tests: name each case as it starts, so a crash in CI identifies itself.
The macOS runner killed the suite process 30 seconds in and the log held
nothing: stdout is block-buffered when it is not a terminal, so the header
printed at startup was still sitting in the buffer when the process died, and
"Subprocess killed" was the entire diagnosis. The runner writes each case name
flushed before running it, which turns that into a name. CTest only prints a
test's output when it fails, so a green run stays exactly as quiet as before.

Also the four warnings Clang reported and MSVC did not, all in dead code that
had accumulated in the cases rather than in anything the framework does: a loop
index compared across signedness, and three sets of energy accumulators that
were computed and never read. Reverb_damping_effectiveness had a whole probe-
tone correlation loop feeding two variables nothing consumed, under a comment
admitting the case "just verifies the reverb produces output" - the spectral
tilt it was meant to measure is already covered properly by the per-band T60
cases, so the loop is gone and the comment now says what the case actually pins.

+++ tests/TestCoreSpectral.cpp
@@ -123,7 +123,7 @@ DSPARK_TEST(FFTReal_silence_is_zero)
     std::vector<float> freq(fft.getFrequencyDomainSize());
     fft.forward(input.data(), freq.data());
 
-    for (int i = 0; i < fft.getFrequencyDomainSize(); ++i)
+    for (size_t i = 0; i < fft.getFrequencyDomainSize(); ++i)
         EXPECT_NEAR(freq[i], 0.0f, 1e-10f);
 }
